Addressing CR feedback
