+++ Assets/HoloToolkit-Examples/SpatialMapping/Scenes/TapToPlace.unity
@@ -88,6 +88,7 @@ LightmapSettings:
     m_PVRFilteringAtrousPositionSigmaDirect: 0.5
     m_PVRFilteringAtrousPositionSigmaIndirect: 2
     m_PVRFilteringAtrousPositionSigmaAO: 1
+    m_ShowResolutionOverlay: 1
   m_LightingDataAsset: {fileID: 0}
   m_UseShadowmask: 1
 --- !u!196 &4
@@ -123,8 +124,8 @@ GameObject:
   - component: {fileID: 181528118}
   - component: {fileID: 181528117}
   - component: {fileID: 181528116}
-  - component: {fileID: 181528115}
   - component: {fileID: 181528114}
+  - component: {fileID: 181528115}
   m_Layer: 1
   m_Name: Cube (2)
   m_TagString: Untagged
@@ -173,7 +174,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 181528112}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dbad991a9894486418e4988222483ef8, type: 3}
+  m_Script: {fileID: 11500000, guid: 12f64a7e6f461aa409b8322f413f96b3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   DefaultGazeDistance: 2
@@ -289,8 +290,8 @@ GameObject:
   - component: {fileID: 532765829}
   - component: {fileID: 532765828}
   - component: {fileID: 532765827}
-  - component: {fileID: 532765826}
   - component: {fileID: 532765825}
+  - component: {fileID: 532765826}
   m_Layer: 4
   m_Name: Cube (1)
   m_TagString: Untagged
@@ -339,11 +340,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 532765823}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dbad991a9894486418e4988222483ef8, type: 3}
+  m_Script: {fileID: 11500000, guid: 12f64a7e6f461aa409b8322f413f96b3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   DefaultGazeDistance: 2
-  PlaceParentOnTap: 1
+  PlaceParentOnTap: 0
   ParentGameObjectToPlace: {fileID: 0}
   IsBeingPlaced: 0
   AllowMeshVisualizationControl: 1
@@ -564,8 +565,8 @@ GameObject:
   - component: {fileID: 1023018530}
   - component: {fileID: 1023018529}
   - component: {fileID: 1023018528}
-  - component: {fileID: 1023018527}
   - component: {fileID: 1023018532}
+  - component: {fileID: 1023018527}
   m_Layer: 0
   m_Name: Cube
   m_TagString: Untagged
@@ -581,7 +582,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1023018526}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dbad991a9894486418e4988222483ef8, type: 3}
+  m_Script: {fileID: 11500000, guid: 12f64a7e6f461aa409b8322f413f96b3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   DefaultGazeDistance: 2
@@ -710,6 +711,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 906323c940a3fad4f8f7e9e4fcd747f4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  containerObject: {fileID: 0}
+  alignmentType: 0
+  stationarySpaceTypePosition: {x: 0, y: 0, z: 0}
+  roomScaleSpaceTypePosition: {x: 0, y: 0, z: 0}
 --- !u!4 &2053883040
 Transform:
   m_ObjectHideFlags: 0
