+++ Assets/HoloToolkit-Examples/SpatialMapping/Scenes/TapToPlace.unity.meta
