+++ Assets/HoloToolkit/Common/Scripts/WorldAnchorManager.cs
@@ -227,81 +227,7 @@ namespace HoloToolkit.Unity
 
             return anchorName;
 #endif
-        }
-
-        /// <summary>
-        /// Attaches an anchor immediately to the specified GameObject.  
-        /// If the anchor store has an anchor with the specified name it will load the anchor, 
-        /// otherwise a new anchor will be saved under the specified name.  
-        /// If no anchor name is provided, the name of the anchor will be the same as the GameObject.
-        /// </summary>
-        /// <param name="gameObjectToAnchor">The GameObject to attach the anchor to.</param>
-        /// <param name="anchorName">Name of the anchor.  If none provided, the name of the GameObject will be used.</param>
-        /// <returns>The the newly attached anchor.</returns>
-        public UnityEngine.XR.WSA.WorldAnchor AttachAnchorImmediate(GameObject gameObjectToAnchor, string anchorName = null)
-        {
-#if !UNITY_WSA || UNITY_EDITOR
-            Debug.LogWarning("World Anchor Manager does not work for this build. AttachAnchorImmediate will not be called.");
-            return null;
-#else
-            if (gameObjectToAnchor == null)
-            {
-                Debug.LogError("[WorldAnchorManager] Must pass in a valid gameObject");
-                return null;
-            }
-
-            // This case is unexpected, but just in case.
-            if (AnchorStore == null)
-            {
-                Debug.LogWarning("[WorldAnchorManager] AttachAnchor called before anchor store is ready.");
-            }
-
-            anchorName = GenerateAnchorName(gameObjectToAnchor, anchorName);
-
-            DoAnchorOperation(
-                new AnchorAttachmentInfo
-                {
-                    AnchoredGameObject = gameObjectToAnchor,
-                    AnchorName = anchorName,
-                    Operation = AnchorOperation.Save
-                });
-
-            return gameObjectToAnchor.GetComponent<WorldAnchor>();
-#endif
-        }
-
-        /// <summary>
-        /// Removes the anchor from the game object immediately and deletes the anchor
-        /// from the anchor store.
-        /// </summary>
-        /// <param name="anchorName">Name of the anchor to remove from the anchor store.</param>
-        /// <param name="gameObjectToUnanchor">GameObject to remove the anchor from.</param>
-        private void RemoveAnchorImmediate(string anchorName, GameObject gameObjectToUnanchor)
-        {
-            if (string.IsNullOrEmpty(anchorName) && gameObjectToUnanchor == null)
-            {
-                Debug.LogWarning("Invalid Remove Anchor Request!");
-                return;
-            }
-
-#if !UNITY_WSA || UNITY_EDITOR
-            Debug.LogWarning("[World Anchor Manager] Does not work for this build. RemoveAnchor will not be called.");
-#else
-            // This case is unexpected, but just in case.
-            if (AnchorStore == null)
-            {
-                Debug.LogWarning("[WorldAnchorManager] RemoveAnchor called before anchor store is ready.");
-            }
-
-            DoAnchorOperation(
-                new AnchorAttachmentInfo
-                {
-                    AnchoredGameObject = gameObjectToUnanchor,
-                    AnchorName = anchorName,
-                    Operation = AnchorOperation.Delete
-                });
-#endif
-        }
+        }        
 
         /// <summary>
         /// Removes the anchor component from the GameObject and deletes the anchor from the anchor store.
