--- Assets/HoloToolkit/Input/Scripts/Utilities/Interactions/TapToPlace.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: dbad991a9894486418e4988222483ef8
-timeCreated: 1463001318
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
