+++ Assets/HoloToolkit/SpatialMapping/Scripts/TapToPlace.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 77b0992d4e37b69498516739501a6a4b
-timeCreated: 1512421196
+guid: 12f64a7e6f461aa409b8322f413f96b3
+timeCreated: 1514933776
 licenseType: Pro
 MonoImporter:
   externalObjects: {}
